# PERF: don't call RangeIndex._data unnecessarily