--- a/pandas/core/indexes/range.py
+++ b/pandas/core/indexes/range.py
@@ -188,6 +188,13 @@ class RangeIndex(Int64Index):
 
     @property
     def _data(self):
+        """
+        An int array that for performance reasons is created only when needed.
+
+        The constructed array is saved in ``_cached_data``. This allows us to
+        check if the array has been created without accessing ``_data`` and
+        triggering the construction.
+        """
         if self._cached_data is None:
             self._cached_data = np.arange(self._start, self._stop, self._step,
                                           dtype=np.int64)