@pep8speaks 2019-05-29
Hello @topper-123! Thanks for updating this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:



* In the file [`pandas/core/indexes/range.py`](https://github.com/pandas-dev/pandas/blob/a5cad779c1b3038abeec3bdff2c1030ef8849c28/pandas/core/indexes/range.py):

> [Line 312:17](https://github.com/pandas-dev/pandas/blob/a5cad779c1b3038abeec3bdff2c1030ef8849c28/pandas/core/indexes/range.py#L312): [E117](https://duckduckgo.com/?q=pep8%20E117) over-indented